+++ packages/react-learn-card/src/components/VCCard/VCCard.tsx
@@ -3,14 +3,42 @@ import { initLearnCard } from '@learncard/core';
 import { VC, Profile, VerificationItem } from '@learncard/types';
 
 import { VCDisplayCard } from '../VCDisplayCard';
+import {
+    VCDisplayCard2,
+    MediaMetadata,
+    VideoMetadata,
+    CredentialIconType,
+} from '../VCDisplayCard2';
 
 export type VCCardProps = {
     credential: VC;
     issueeOverride?: Profile;
     className?: string;
+    version?: '1' | '2';
+
+    /* Only used for version 2 */
+    subjectImageComponent?: React.ReactNode;
+    // convertTagsToSkills?: (tags: string[]) => { [skill: string]: string[] };
+    handleXClick?: () => void;
+    getFileMetadata?: (url: string) => MediaMetadata;
+    getVideoMetadata?: (url: string) => VideoMetadata;
+    onMediaAttachmentClick?: (url: string) => void;
+    bottomRightIcon?: CredentialIconType;
 };
 
-export const VCCard: React.FC<VCCardProps> = ({ credential, issueeOverride, className = '' }) => {
+export const VCCard: React.FC<VCCardProps> = ({
+    credential,
+    issueeOverride,
+    className = '',
+    version = '1',
+    subjectImageComponent,
+    // convertTagsToSkills,
+    handleXClick,
+    getFileMetadata,
+    getVideoMetadata,
+    onMediaAttachmentClick,
+    bottomRightIcon,
+}) => {
     const [loading, setLoading] = useState(true);
     const [vcVerification, setVCVerification] = useState<VerificationItem[]>([]);
 
@@ -25,13 +53,30 @@ export const VCCard: React.FC<VCCardProps> = ({ credential, issueeOverride, clas
         verify();
     }, [credential]);
 
+    if (version === '1') {
+        return (
+            <VCDisplayCard
+                credential={credential}
+                issueeOverride={issueeOverride}
+                className={className}
+                loading={loading}
+                verification={vcVerification}
+            />
+        );
+    }
     return (
-        <VCDisplayCard
+        <VCDisplayCard2
             credential={credential}
             issueeOverride={issueeOverride}
-            className={className}
-            loading={loading}
-            verification={vcVerification}
+            verificationInProgress={loading}
+            verificationItems={vcVerification}
+            subjectImageComponent={subjectImageComponent}
+            // convertTagsToSkills={convertTagsToSkills}
+            handleXClick={handleXClick}
+            getFileMetadata={getFileMetadata}
+            getVideoMetadata={getVideoMetadata}
+            onMediaAttachmentClick={onMediaAttachmentClick}
+            bottomRightIcon={bottomRightIcon}
         />
     );
 };