+++ packages/react-learn-card/src/components/VCVerificationCheck/VCVerificationCheck.tsx
@@ -14,7 +14,7 @@ export const VCVerificationCheck: React.FC<VCVerificationCheckProps> = ({ size =
     };
     return (
         <div className="vc-verification-wrapper" style={{ position: 'relative' }}>
-            <div className="flex items-center justify-center rounded-[50%] bg-white rounded-full overflow-hidden">
+            <div className="flex items-center justify-center bg-white rounded-full overflow-hidden">
                 <img
                     className="h-full w-full object-contain p-1"
                     src={VerificationCheckDisplay ?? ''}
@@ -28,19 +28,44 @@ export const VCVerificationCheck: React.FC<VCVerificationCheckProps> = ({ size =
 
 export default VCVerificationCheck;
 
-export const VCVerificationCheckWithText: React.FC<VCVerificationCheckProps> = ({
-    size = '60px',
-    loading = true,
-}) => {
-    const [defaultLoading, setDefaultLoading] = useState(true);
+export const VCVerificationCheckWithSpinner: React.FC<
+    VCVerificationCheckProps & { spinnerSize?: string; withBorder?: boolean }
+> = ({ size = '44px', spinnerSize = '60px', loading = true, withBorder = false }) => {
+    const loadingState = loading ? CircleLoadingState.spin : CircleLoadingState.stop;
 
     const spinnerStyle: React.CSSProperties = {
         position: 'absolute',
         top: 0,
         left: 0,
-        height: '60px', //todo calculate this based on other
-        width: '60px',
+        height: spinnerSize,
+        width: spinnerSize,
     };
+
+    const spinnerSizeNumber = parseInt(spinnerSize.replace('px', ''));
+    const offset = spinnerSizeNumber > 50 ? 3 : 2;
+
+    return (
+        <div
+            style={{ width: spinnerSize, height: spinnerSize, position: 'relative' }}
+            className={`flex items-center justify-center bg-white rounded-full overflow-hidden ${withBorder ? ' ml-3 mr-3' : ''
+                }`}
+        >
+            <VCVerificationCheck size={size} loading={loading} />
+            <div className="vc-verification-spinner-overlay" style={spinnerStyle}>
+                <CircleSpinner
+                    loadingState={loadingState}
+                    size={spinnerSizeNumber}
+                    marginOffset={offset}
+                />
+            </div>
+        </div>
+    );
+};
+
+export const VCVerificationCheckWithText: React.FC<
+    VCVerificationCheckProps & { spinnerSize?: string }
+> = ({ size = '44px', spinnerSize = '60px', loading = true }) => {
+    const [defaultLoading, setDefaultLoading] = useState(true);
 
     useEffect(() => {
         let timer1 = setTimeout(() => setDefaultLoading(false), 3 * 1000);
@@ -49,21 +74,15 @@ export const VCVerificationCheckWithText: React.FC<VCVerificationCheckProps> = (
         };
     });
 
-    const loadingState =
-        loading || defaultLoading ? CircleLoadingState.spin : CircleLoadingState.stop;
-
     return (
         <div className="flex justify-center items-center translate-x-[10px] w-full vc-verification-full-wrapper">
             <span className="text-white font-bold tracking-wider">Verified</span>
-            <div
-                style={{ width: size, height: size, position: 'relative' }}
-                className={`flex items-center justify-center bg-white rounded-full ml-3 mr-3 overflow-hidden`}
-            >
-                <VCVerificationCheck size={'44px'} loading={loading} />
-                <div className="vc-verification-spinner-overlay" style={spinnerStyle}>
-                    <CircleSpinner loadingState={loadingState} />
-                </div>
-            </div>
+            <VCVerificationCheckWithSpinner
+                size={size}
+                spinnerSize={spinnerSize}
+                loading={loading || defaultLoading}
+                withBorder
+            />
             <span className="text-white font-bold tracking-wider">Credential</span>
         </div>
     );