+++ packages/react-learn-card/src/constants/icons.ts
@@ -4,6 +4,7 @@ import trophy from '../assets/images/trophy.svg';
 import briefcase from '../assets/images/briefcase.svg';
 import graduation from '../assets/images/graduation.svg';
 import lightbulb from '../assets/images/lightbulb.svg';
+import award from '../assets/images/award.svg';
 
 import puzzlelight from '../assets/images/minipuzzle.svg';
 import trophylight from '../assets/images/minitrophy.svg';
@@ -20,6 +21,7 @@ export {
     gradcaplight,
     trophylight,
     puzzlelight,
+    award,
 };
 
 export const ICONS_TO_SOURCE = {
@@ -32,6 +34,7 @@ export const ICONS_TO_SOURCE = {
     [Icons.puzzlelight]: puzzlelight,
     [Icons.trophylight]: trophylight,
     [Icons.gradcaplight]: gradcaplight,
+    [Icons.award]: award,
 };
 
 export const TYPE_TO_ICON: any = {