+++ packages/react-learn-card/src/constants/index.ts
@@ -4,7 +4,7 @@ import jobhistoryGraphic from '../assets/images/walletjobhistory.webp';
 import learningHistoryGraphic from '../assets/images/walletlearninghistory.webp';
 import skillsGraphic from '../assets/images/walletskills.webp';
 import achievementsGraphic from '../assets/images/walletTrophy.webp';
-
+import socialBadge from '../assets/images/social-badge.png';
 
 import MiniTrophyIcon from '../assets/images/minitrophycolored.svg';
 import MiniGradIcon from '../assets/images/minigradcapcolored.svg';
@@ -24,6 +24,7 @@ export const TYPE_TO_IMG_SRC = {
     [WalletCategoryTypes.learningHistory]: learningHistoryGraphic,
     [WalletCategoryTypes.currency]: coinsGraphic,
     [WalletCategoryTypes.skills]: skillsGraphic,
+    [WalletCategoryTypes.socialBadge]: socialBadge,
 };
 
 export const TYPE_TO_WALLET_COLOR = {
@@ -33,6 +34,7 @@ export const TYPE_TO_WALLET_COLOR = {
     [WalletCategoryTypes.learningHistory]: 'emerald-300',
     [WalletCategoryTypes.currency]: 'cyan-200',
     [WalletCategoryTypes.skills]: 'indigo-300',
+    [WalletCategoryTypes.socialBadge]: 'cyan-200',
 };
 
 export const TYPE_TO_WALLET_LIGHT_COLOR = {
@@ -42,6 +44,7 @@ export const TYPE_TO_WALLET_LIGHT_COLOR = {
     [WalletCategoryTypes.learningHistory]: 'emerald-200',
     [WalletCategoryTypes.currency]: 'cyan-200',
     [WalletCategoryTypes.skills]: 'indigo-200',
+    [WalletCategoryTypes.socialBadge]: 'cyan-200',
 };
 
 export const TYPE_TO_WALLET_DARK_COLOR = {
@@ -51,4 +54,5 @@ export const TYPE_TO_WALLET_DARK_COLOR = {
     [WalletCategoryTypes.learningHistory]: 'emerald-500',
     [WalletCategoryTypes.currency]: 'cyan-200',
     [WalletCategoryTypes.skills]: 'indigo-400',
+    [WalletCategoryTypes.socialBadge]: 'cyan-200',
 };