+++ packages/react-learn-card/src/types/index.ts
@@ -1,4 +1,9 @@
-import { VerificationItem, VC, Profile, CredentialSubject } from '@learncard/types';
+import {
+    VerificationItem,
+    AchievementCredential,
+    Profile,
+    CredentialSubject,
+} from '@learncard/types';
 import React from 'react';
 
 export enum Icons {
@@ -11,6 +16,7 @@ export enum Icons {
     gradcaplight,
     trophylight,
     puzzlelight,
+    award,
 }
 
 export enum LCSubtypes {
@@ -28,10 +34,12 @@ export enum WalletCategoryTypes {
     currency = 'currency',
     learningHistory = 'learningHistory',
     skills = 'skills',
+    socialBadge = 'socialBadge',
 }
 
 export type CredentialInfo = {
     title?: string;
+    imageUrl?: string;
     createdAt?: string;
     issuer?: Profile;
     issuee?: Profile;
@@ -235,4 +243,16 @@ export type GenericCardProps = {
     checkStatus?: boolean;
     onClick?: () => void;
     flipped?: boolean;
+};
+
+export type Attachment = {
+    title: string;
+    url: string;
+    type: 'photo' | 'document' | 'video' | 'link';
+};
+
+export type BoostAchievementCredential = AchievementCredential & {
+    display?: { backgroundImage?: string; backgroundColor?: string };
+    image: string;
+    attachments: Attachment[];
 };